Fix bug on browser detection logic (Thanks to @TCashion)
This bug catch and fix are thanks to @TCashion on PR#52 

When browser it's defined as id on HTML element cause a false positive detecting browser as Chrome and make module stop working on other navigators (Firefox for example).

We switch to detect Firefox looking for `InstallTrigger` interface, that it's present on Firefox and non standard, so we can use it safely to detect Firefox browsers. (More info on: https://developer.mozilla.org/en-US/docs/Web/API/InstallTrigger).